+++ src/domain/Models/Salary.cs
@@ -5,8 +5,8 @@ namespace GovUk.Education.SearchAndCompare.Domain.Models
 {
     public class Salary
     {
-        public long? Minimum { get; set; }
+        public int? Minimum { get; set; }
 
-        public long? Maximum { get; set; }
+        public int? Maximum { get; set; }
     }
 }
